Sagas now split for the 2 pages
