+++ gui/app/queue-admin.js
@@ -2,15 +2,12 @@ import regeneratorRuntime from "regenerator-runtime/runtime";
 const React = require("react");
 const ReactDOM = require("react-dom");
 import { AppContainer } from "react-hot-loader";
-// Main app
 import QueueAdminGUI from "./components/queue-admin-gui";
-// Redux related
 import { Provider } from "react-redux";
 import configureStore from "./reducers/admin_queue_store";
-
-// Webpack will bundle styling
 import "bootstrap/dist/css/bootstrap.min.css";
 import "font-awesome/css/font-awesome.min.css";
+import { pullQueueRules } from "./actions/queues";
 
 let store = configureStore();
 
@@ -26,6 +23,8 @@ const render = Component => {
 };
 
 render(QueueAdminGUI);
+// Initial loading of queue rules to populate UI with data
+store.dispatch(pullQueueRules());
 
 // Webpack Hot Module Replacement API
 if (module.hot) {
