+++ gui/app/reducers/docstore_gui_store.js
@@ -1,4 +1,6 @@
 import docstoreGuiRootReducer from "./index";
 import config from "./configure_store";
+import rootSaga from "../sagas/docstore_gui_sagas";
 
-export default initialState => config(docstoreGuiRootReducer, initialState);
+export default initialState =>
+  config(docstoreGuiRootReducer, rootSaga, initialState);
