+++ gui/app/sagas/admin_queue_sagas.js
@@ -1,0 +1,8 @@
+import { call, fork, takeLatest, select, put, all } from "redux-saga/effects";
+import * as queueActions from "../actions/queues";
+
+export default function* root() {
+  yield all([
+    //fork(watchSearch),
+  ]);
+}
