+++ gui/test/search_sagas.test.js
@@ -1,10 +1,10 @@
 import configureStore from "../app/reducers/docstore_gui_store";
 import { takeLatest, call, put, select } from "redux-saga/effects";
-import {
+import ourSaga, {
   watchSearch,
   fetchBibliographicPost,
   getSearchParameter
-} from "../app/sagas/index";
+} from "../app/sagas/docstore_gui_sagas";
 import {
   SEARCH_BIB_RECORD_ID,
   searchBibRecord,
@@ -13,7 +13,6 @@ import {
 import * as searchActions from "../app/actions/searching";
 import SagaTester from "redux-saga-tester";
 import reducers from "../app/reducers/docstore_gui_root_reducer";
-import ourSaga from "../app/sagas";
 
 import api from "../app/api/index";
